Create personalvoiceassistantv1
It detects your sound and turns into text, meanwhile activates some functions including control of your browser to search something on google or open the youtube. The duration of audio processing and response time is dependent on the libraries. Language is in turkish for this project. It can do search on google or play a multiplication game or tell you the time.